app/requirements.txt
astroid==3.1.0
cfgv==3.4.0
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.14.0
identify==2.5.36
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
pylint==3.1.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
virtualenv==20.26.1
